fix soft_version partition for archer c25
Flashing LEDE from the original firmware's GUI, a version string taken
from the soft_ver partition of the firmware image is written to the
router's config partition.

When using tftp recovery to go back to the original Archer C25 firmware,
a version check compares that version to the version of the firmware to
be flashed.

Without proper contents in the config partition, reverting to the
original firmware fails.

Therefore, write the string "soft_ver:1.0.0\n" to that soft_ver
partition.